Started getting settings page functioanlity working
Started creation of a sidebar that is commonly found in modern web apps, working on creation of this with corresponding components in order to have page function as intended. As of now, just have boilerplate for account and privacy settings. no actual stuff is done here yet, but will be in the future. Also reworked settings page to accomodate need of useState functions. Aside from this, made some other small changes