Settings page stays live
Got it so the settings page no longer bricks itself if a steam user is logged in, and now does proper checks for if the user is a normal user or a steam user. This does not change the overall functionality of the site as that still needs to be adjusted and fixed in order to accommodate steam users getting their steam information. However, this is a good step forward